Use appropriate units for file I/O
The MPAS framework subroutines mpas_new_unit and mpas_release_unit ensure that file units selected for I/O do not conflict with other units already used throughout MPAS or in a surroundingprogram when MPAS is used as a library.  These changes to core_atmosphere ensures file units are selected appropriately.  The update is dependent on an earlier change to the framework that selects units for unformatted binary I/O from a small subset.