Remove YSU theta -> t -> theta tendency computations
At the end of the ysu subroutine, this mod removes the computation
that switches the temperature tendency to a potential temperature
tendency.

We make that change because inside of the bl_ysu_run subroutine, we
removed the computation that changed the potential temperature
tendency to a temperature tendency.

At the end of the bl_ysu_run routine, we have a potential temperature
tendency. There is no need to temporarily convert this to a temperature
tendency, only to reconvert it back to a potential temperature
tendency immediately after the routine exits.

As expected, this change causes bit-for-bit differences. This change
was not included in the sequence of ysu mods that give bit-wise identical
results so as to allow the far larger code modifications to be validated
more easily.

Both MPAS and WRF physics developers OK'ed this modification. Both
MPAS and WRF physics schemes expect to handle potential temperature
tendencies coming out of the PBL scheme from phyics.

Advantages:
1. No need to do reciprocal computations to get back to the same
state variable (algebraically the same).
2. The trailing computation in the ysu scheme (which is now removed)
would have required an extra CPF interstitial routine.

modified:   src/core_atmosphere/physics/physics_sima/bl_ysu.F