Read full initial state from two streams, "static" and "ICs"
Rather than reading the initial model state from just the "input" stream,
time-invariant fields are now read from the "static" stream and all
time-varying fields are read from the "ICs" stream.

Since the time-invariant fields might be stored in a file with a time
stamp different from the start time of the model, which is not a problem,
the MPAS_STREAM_NEAREST whence argument to mpas_stream_mgr_read has been
added for the "static" stream.