Add a class to help processing static geogrid fields
To help with future efforts in parallelizing the interpolation of static
fields, a class has been added, mpas_geotile_type, which helps with reading and
storing static datasets. As well, it contains routines for converting pixel
coordinates into latitude and longitude.

Currently, the mpas_geotile_type is not being used anywhere within the
init_atmosphere core, it is only being compiled.